Long overflow bug fixed.
Setting initial timeOut=Long.Max_Value/2 causes overflow.

+++ src/edu/mit/streamjit/impl/distributed/TailChannels.java
@@ -266,7 +266,7 @@ public class TailChannels {
 		 * <code>time=unnormalizedTime(normalizedTime(time))</code>
 		 */
 		protected long unnormalizedTime(long time) {
-			return (time * (totalCount - skipCount)) / Options.outputCount;
+			return (long) (time * (((double) (totalCount - skipCount)) / Options.outputCount));
 		}
 
 		/**
